move dropdownAutoComplete into it's own component
